refactor(admin): render the activity feed and history timelines as lists
All three registered a TanStack column helper, features, useTable and
FlexRender, then rendered a <ul>. The activity feed had zero features;
the timelines existed for one newest/oldest toggle. A map and a
useState do the same job, and the shared Timeline holds the toggle.
